+++ packages/bodiless-core/src/PageEditContext/types.ts
@@ -13,21 +13,8 @@
  */
 
 // Internal mobx store which holds the state.
-import React from 'react';
 import { TOverlaySettings } from '../Types/PageOverlayTypes';
-
-export type TMenuOption = {
-  name: string;
-  icon?: (() => string) | string;
-  label?: (() => string) | string;
-  isActive?: (() => boolean) | boolean;
-  isDisabled?: (() => boolean) | boolean;
-  isHidden?: (() => boolean) | boolean;
-  handler?: (event: React.MouseEvent) => any;
-  local?: boolean;
-  global?: boolean;
-  group?: string;
-};
+import type { TMenuOption } from '../Types/ContextMenuTypes';
 
 export type TMenuOptionGetter = () => TMenuOption[];
