+++ packages/bodiless-richtext/src/asSlateCustomComponent.tsx
@@ -24,6 +24,7 @@ const asSlateCustomComponent = <P extends WithNodeProps>(Component: CT<P>) => ({
   const getters = {
     getNode: () => componentData,
     getKeys: () => ['slatenode'],
+    hasError: () => false,
   };
   const actions = {
     // tslint: disable-next-line:no-unused-vars
